Major refactoring of many of the traits in MapReduceSugar
This should make the code a bit more modular and simple. It also solves
some of the problems with the previous implementation having trouble
with things such as running a simple partition job with no mapper or
reducer being set.
Addition of more integration tests and some unit tests for Hadoop<->Scala conversions